+++ src/renderer/pages/Staking/Overview/common/types.ts
@@ -1,0 +1,11 @@
+import { Account, Address, Unlocking } from '@/src/renderer/shared/core';
+
+export type NominatorInfo<T = Account> = {
+  address: Address;
+  isSelected: boolean;
+  stash?: Address;
+  account: T;
+  totalReward?: string;
+  totalStake?: string;
+  unlocking?: Unlocking[];
+};
