+++ src/renderer/pages/Staking/Overview/components/NominatorsList/NominatorsList.tsx
@@ -4,12 +4,13 @@ import { Trans } from 'react-i18next';
 import { useI18n } from '@renderer/app/providers';
 import { cnTw } from '@shared/lib/utils';
 import { FootnoteText, Tooltip, Icon, HelpText } from '@shared/ui';
-import type { Asset, Explorer, Address, Account, NominatorInfo, ShardAccount } from '@shared/core';
+import type { Asset, Explorer, Address, Account, ShardAccount } from '@shared/core';
 import { useStakingData } from '@entities/staking';
 import { AccountAddress, AddressWithName, accountUtils } from '@entities/wallet';
 import { NominatorsItem } from './NominatorItem/NominatorItem';
 import { ShardedList } from './ShardedList/ShardedList';
 import { TimeToEra } from '../TimeToEra/TimeToEra';
+import { NominatorInfo } from '../../common/types';
 
 type Props = {
   nominators: (NominatorInfo<ShardAccount>[] | NominatorInfo)[];
