+++ src/renderer/shared/core/index.ts
@@ -46,4 +46,4 @@ export type { Identity, SubIdentity } from './types/identity';
 export type { Validator } from './types/validator';
 
 export { RewardsDestination } from './types/stake';
-export type { Stake, Unlocking, NominatorInfo } from './types/stake';
+export type { Stake, Unlocking } from './types/stake';
