+++ src/renderer/shared/core/types/stake.ts
@@ -1,4 +1,3 @@
-import { Account } from './account';
 import type { Address, ChainId } from './general';
 
 export type Stake = {
@@ -20,13 +19,3 @@ export const enum RewardsDestination {
   RESTAKE = 'restake',
   TRANSFERABLE = 'transferable',
 }
-
-export type NominatorInfo<T = Account> = {
-  address: Address;
-  isSelected: boolean;
-  stash?: Address;
-  account: T;
-  totalReward?: string;
-  totalStake?: string;
-  unlocking?: Unlocking[];
-};
